+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 ### Added
-- When a symbol is incorrectly defined multiple times, diagnostics show the locations in the hover box.
+- When a symbol or message is incorrectly defined multiple times, diagnostics show the locations in the hover box.
 
 ### Improved
 - Improved performance of diagnostics.